Fix compiler warnings: CMake, vm_x86, tr_bsp, tr_cmds, tr_flares, tr_image, sdl_input, newlines
- CMake: remove duplicate RENDERER_DEFAULT from client target (already global)
- vm_x86: DROP() variadic macro zero-arg fix; emit_modrm_offset in #if !idx64
- tr_bsp: ParseFlare unused params; sign-compare for contents/brushesCount/sidesCount/size
- tr_cmds, tr_flares, tr_image: sign-compare casts (size_t, unsigned)
- sdl_input: joystick button count vs ARRAY_LEN sign-compare
- tr_shade_calc, vk_mikktspace: newline at EOF

+++ src/renderers/opengl/tr_cmds.c
@@ -131,7 +131,7 @@ static void *R_GetCommandBufferReserved( int bytes, int reservedBytes ) {
 
 	// always leave room for the end of list command
 	if ( cmdList->used + bytes + sizeof( int ) + reservedBytes > MAX_RENDER_COMMANDS ) {
-		if ( bytes > MAX_RENDER_COMMANDS - sizeof( int ) ) {
+		if ( (size_t)bytes > MAX_RENDER_COMMANDS - sizeof( int ) ) {
 			ri.Error( ERR_FATAL, "R_GetCommandBuffer: bad size %i", bytes );
 		}
 		// if we run out of room, just start dropping commands
